Refactor: Python APIの変換系を移動
